requirements.txt
flask
pillow
numpy
requests
gunicorn